+++ sudoku.c
@@ -6,33 +6,33 @@
 #include <unistd.h>
 #include <math.h>
 
-int get(Sudoku *sudoku, int x, int y)
+size_t get(Sudoku *sudoku, size_t x, size_t y)
 {
     return sudoku->grid[x * sudoku->edge + y];
 }
 
-void put(Sudoku *sudoku, int x, int y, int value)
+void put(Sudoku *sudoku, size_t x, size_t y, size_t value)
 {
     sudoku->grid[x * sudoku->edge + y] = value;
 }
 
 void print_sudoku(Sudoku *sudoku)
 {
-    for (int i = 0; i < sudoku->edge; i++)
+    for (size_t i = 0; i < sudoku->edge; i++)
     {
-        for (int j = 0; j < sudoku->edge; j++)
-            printf("%2d", get(sudoku, i, j));
+        for (size_t j = 0; j < sudoku->edge; j++)
+            printf("%2ld", get(sudoku, i, j));
         printf("\n");
     }    
     printf("\n");
 }
 
-Sudoku *get_sudoku_from_file(char *pathname, int size)
+Sudoku *get_sudoku_from_file(char *pathname, size_t size)
 {
-    int fd = open(pathname, O_RDONLY);
+    size_t fd = open(pathname, O_RDONLY);
 
     char *content = malloc(size * 2 * sizeof(char));
-    int reading_result = read(fd, content, size * 2);
+    size_t reading_result = read(fd, content, size * 2);
     if (reading_result != size * 2)
         fprintf(stderr, "Error while reading the file or the grid is not complete.\n");
 
@@ -44,8 +44,8 @@ Sudoku *get_sudoku_from_file(char *pathname, int size)
     sudoku->edge = sqrt(size);
     sudoku->grid = malloc(size * sizeof(int));
 
-    int i = 0;
-    int y = 0;
+    size_t i = 0;
+    size_t y = 0;
     while (i < size && y < size * 2)
     {
         if (content[y] == '\n' || content[y] == ' ')
